# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Routine Builder AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/messages:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a chat message (SSE streaming)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: SSE stream of response chunks
          content:
            text/event-stream:
              schema:
                type: string
...
components:
  schemas:
...
    ChatMessageInput:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: The user's message
        previousResponseId:
          type: ["string", "null"]
          description: The previous OpenAI response ID for conversation chaining
        uploadedFileId:
          type: ["string", "null"]
          description: Optional OpenAI file ID to attach to this message
...