# lib/api-zod/src/generated/types/chatMessageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Routine Builder AI API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatMessageInput {
  /** The user's message */
  message: string;
  /**
     * The previous OpenAI response ID for conversation chaining
     * @nullable
     */
  previousResponseId?: string | null;
  /**
     * Optional OpenAI file ID to attach to this message
     * @nullable
     */
  uploadedFileId?: string | null;
}